Fill NA with 0 (For trainings of only certain type) (#15)
I want to begin doing specific training sessions with specific poor performing types (Such as residential). So this means that the jobs.json may not contain a certain type (Such as not including Commercial or Industrial) if I am only training on residential. Before it would throw an error since the data would come across as NA instead of 0.